update: refactor DefaultLayout
