--- packages/ui-components/src/Templates/defaultLayout/index.ts
@@ -1,1 +1,0 @@
-export * from "./DefaultLayout";
