--- packages/ui-components/src/Templates/index.ts
@@ -1,1 +1,0 @@
-export * from "./defaultLayout";
