+++ packages/ui-components/src/components/Organisms/index.ts
@@ -1,9 +1,8 @@
+export * from './BreadCrumbs'
+export * from './Footer'
+export * from './SideNav'
+export * from './UploadButton'
 export * from './modals'
 export * from './steps'
 export * from './table'
-export * from './BottomNav'
-export * from './NavBar'
-export * from './SideNav'
-export * from './TopBar'
-export * from './UploadButton'
-export * from './BreadCrumbs'
+
