+++ packages/ui-components/src/index.tsx
@@ -1,6 +1,6 @@
 import "./styles.css";
 
-export * from "./Templates";
+// export * from "./Templates";
 export * from "./components";
 export * from "./hooks/useScreen";
 export * from "./icons";
