+++ packages/ui-components/tsconfig.json
@@ -1,6 +1,6 @@
 {
   // see https://www.typescriptlang.org/tsconfig to better understand tsconfigs
-  "include": ["src", "types", "../web-app/public/animations"],
+  "include": ["src", "types", "../web-app/public/animations", "../web-app/components/TopBar.tsx", "../web-app/components/BottomNav.tsx"],
   "compilerOptions": {
     "module": "esnext",
     "lib": ["dom", "esnext"],
