+++ packages/web-app/components/BottomNav.tsx
@@ -1,6 +1,6 @@
+import Nation3Logo from "@nation3/ui-components/src/components/Atoms/Nation3Logo";
 import Link from "next/link";
 import React, { ReactNode } from "react";
-import Nation3Logo from "../Atoms/Nation3Logo";
 
 type IBottomNavProps = {
 	connectionButton?: ReactNode;
